--- a/web/api/query.go
+++ b/web/api/query.go
@@ -156,7 +156,7 @@ func (serv MetricsService) Metrics(w http.ResponseWriter, r *http.Request) {
 	setAccessControlHeaders(w)
 	w.Header().Set("Content-Type", "application/json")
 
-	metricNames := serv.Storage.GetLabelValuesForLabelName(clientmodel.MetricNameLabel)
+	metricNames := serv.Storage.LabelValuesForLabelName(clientmodel.MetricNameLabel)
 	sort.Sort(metricNames)
 	resultBytes, err := json.Marshal(metricNames)
 	if err != nil {