Finally... I was working in increments on this for an eternity...

```
benchmark                             old ns/op     new ns/op     delta
BenchmarkGetValueAtTimeChunkType1     23514417      3806180       -83.81%
BenchmarkGetValueAtTimeChunkType0     17600036      3945522       -77.58%
BenchmarkGetRangeValuesChunkType1     46013168      12004930      -73.91%
BenchmarkGetRangeValuesChunkType0     36569615      12893927      -64.74%

benchmark                             old allocs     new allocs     delta
BenchmarkGetRangeValuesChunkType1     472130         21893          -95.36%
BenchmarkGetRangeValuesChunkType0     493789         23697          -95.20%
BenchmarkGetValueAtTimeChunkType0     230743         20094          -91.29%
BenchmarkGetValueAtTimeChunkType1     230191         20049          -91.29%

benchmark                             old bytes     new bytes     delta
BenchmarkGetValueAtTimeChunkType1     3844556       485557        -87.37%
BenchmarkGetValueAtTimeChunkType0     3851663       489308        -87.30%
BenchmarkGetRangeValuesChunkType0     14887560      7463797       -49.87%
BenchmarkGetRangeValuesChunkType1     14371182      7255937       -49.51%
```