@fabxc 2015-05-19
Those are great results!

I'll give the code a closer look tomorrow.
Could you give an estimate how those improvements translate to different kinds of queries?