@beorn7 2015-05-19
@fabxc Not sure how much that will show up in real life. Anything that gets a lot from the same series will benefit most. We had a case at SC, where a data-intensive slow query was spending most of the time in the chunk iterators.