@fabxc 2015-05-20
Careful :+1: from me.

LGTM, but I obviously have no full understanding of everything that might affect or be affected by those changes.