@beorn7 2015-05-20
@juliusv What are your thoughts about the get... thing? (You introduced the naming after all...) I'm kind of tempted to rip out all the get's. As said, could be a different change, but my changes to the iterator interface created a lot of new get... methods...