@grobie 2015-05-20
I don't read a get prefix as "will trigger more processing than a method without get". :+1: to remove the prefix everywhere.