@fabxc 2015-05-20
Which also are in the exported Storage interface.

On Wed, May 20, 2015 at 4:36 PM Julius Volz <notifications@github.com>
wrote:

> Screw the gets!
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/prometheus/prometheus/pull/706#issuecomment-103908534>
> .
>